test: extract overlay Esc handling tests into escape.test.js
